+++ CHANGES.rst
@@ -10,7 +10,8 @@ Incompatibilities:
 
 New:
 
-- *add item here*
+- Add ImagesScales vocabulary.
+  [bsuttor]
 
 Fixes: