+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+2.1.25 (unreleased)
+-------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 2.1.24 (2016-04-15)
 -------------------